Update a11y test to have an inaccessible form.
As of axe-core 4.1.0, form inputs with only a placeholder are
considered accessible, so the 'Inaccessible' form component is no
longer inaccessible.

This commit changes the 'Inaccessible' form component to only have an
'id' attribute, and updates the 'inaccessible forms fail axe' test
to expect an accessibility violation, rather than swallowing the error.